Change Harga Tali field to decimal

+++ YkCalculator/YkCalculator.Model/InputField.cs
@@ -124,7 +124,7 @@ namespace YkCalculator.Model
             {
                 DisplayName = "Harga Tali",
                 PropertyName = "HargaTali",
-                PropertyType = "double",
+                PropertyType = "decimal",
                 Required = false
             });
         }
